fix(calm-hub-ui): address review feedback on namespace browse page
- Hub: a failed namespace-counts fetch now flags failure separately so the
  type tabs render resting (no badge) instead of a dimmed 0 that contradicts
  the independently-loaded item grid.
- Derive NamespaceResourceType as Exclude<TypeInUI, 'Controls'> and export a
  shared COUNT_FIELD map; MobileNavMenu uses it instead of a hand switch whose
  silent default would drop counts for any future type.
- Extract useActiveType into its own file (component-decomposition rule).
- Memoize activeItems for consistency with the surrounding memoized values.
